Make tools/cpp code pass Skylint.
This is important because if the code in `tools/` does not run with
`--all_incompatible_changes` turned on, then any downstream Skylark
code in any project that uses Bazel as the build system and uses
`@bazel_tools` will not work with `--all_incompatible_changes` turned
on. This prevents people from e.g. turning this flag on in CI.